+++ src/bin/run_pcm.cpp
@@ -38,121 +38,71 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/format.hpp>
 
-#include "bi_operators/BoundaryIntegralOperator.hpp"
-#include "bi_operators/BIOperatorData.hpp"
-#include "cavity/Cavity.hpp"
-#include "green/IGreensFunction.hpp"
-#include "interface/Input.hpp"
-#include "utils/Molecule.hpp"
-#include "solver/PCMSolver.hpp"
-#include "utils/Atom.hpp"
-#include "Citation.hpp"
-#include "utils/cnpy.hpp"
-#include "utils/Factory.hpp"
-#include "utils/Solvent.hpp"
-#include "utils/Sphere.hpp"
+#include "interface/Meddle.hpp"
 
-//#include "BuildInfo.hpp"
-extern "C" void host_writer(const char * message, PCMSolverIndex message_length);
+std::ofstream pcmsolver_out;
+
+extern "C" void host_writer(const char * message, int message_length);
+
+std::string remove_extension(const std::string & filename);
 
 int main(int argc, char * argv[]) {
-  // Open output file pcmsolver.out
-  std::ofstream out_stream;
-  out_stream.open("pcmsolver.out");
-
   if (argc > 2)
     PCMSOLVER_ERROR("Too many arguments supplied");
+  using namespace pcm;
+
   TIMER_ON("Input parsing");
-  Input parsed(argv[1]);
+  Input input(argv[1]);
   TIMER_OFF("Input parsing");
+  TIMER_ON("Initializing molecule");
+  input.initMolecule();
+  TIMER_OFF("Initializing molecule");
+  Meddle context_(input);
 
-  // Create cavity
-  TIMER_ON("Generating cavity");
-  Cavity * cavity = Factory<Cavity, cavityData>::TheFactory().create(
-      parsed.cavityType(), parsed.cavityParams());
-  TIMER_OFF("Generating cavity");
-  // Create Green's functions
-  // INSIDE
-  TIMER_ON("Generating Green's function inside");
-  IGreensFunction * gf_i = Factory<IGreensFunction, greenData>::TheFactory().create(
-      parsed.greenInsideType(), parsed.insideGreenParams());
-  TIMER_OFF("Generating Green's function inside");
-  // OUTSIDE
-  TIMER_ON("Generating Green's function outside");
-  IGreensFunction * gf_o = Factory<IGreensFunction, greenData>::TheFactory().create(
-      parsed.greenOutsideType(), parsed.outsideStaticGreenParams());
-  TIMER_OFF("Generating Green's function outside");
-  TIMER_ON("Generating solver");
-  PCMSolver * solver = Factory<PCMSolver, solverData>::TheFactory().create(
-      parsed.solverType(), parsed.solverParams());
-  TIMER_OFF("Generating solver");
+  // Prepare output filename
+  pcmsolver_out.open(remove_extension(argv[1]).erase(0, 1) + ".out");
+  context_.printInfo();
 
-  BoundaryIntegralOperator * op =
-      Factory<BoundaryIntegralOperator, biOperatorData>::TheFactory().create(
-          parsed.integratorType(), parsed.integratorParams());
-
-  TIMER_ON("Building system matrix");
-  solver->buildSystemMatrix(*cavity, *gf_i, *gf_o, *op);
-  TIMER_OFF("Building system matrix");
-  // Always save the cavity in a cavity.npz binary file
-  // Cavity should be saved to file in initCavity(), due to the dependencies of
-  // the WaveletCavity on the wavelet solvers it has to be done here...
-  TIMER_ON("Save cavity to .npz file");
-  cavity->saveCavity();
-  TIMER_OFF("Save cavity to .npz file");
-
-  // Printout relevant info
-  out_stream << "~~~~~~~~~~ PCMSolver ~~~~~~~~~~" << std::endl;
-  // How the executable was built
-  // out_stream << buildInfo() << std::endl;
-  // How the calculation was set up
-  out_stream << "Using CODATA " << parsed.CODATAyear() << " set of constants."
-             << std::endl;
-  out_stream << "Input parsing done " << parsed.providedBy() << std::endl;
-  out_stream << "========== Cavity " << std::endl;
-  out_stream << "Atomic radii set: " << parsed.radiiSetName() << std::endl;
-  out_stream << *cavity << std::endl;
-  out_stream << "========== Solver " << std::endl;
-  out_stream << *solver << std::endl;
-  out_stream << "============ Medium " << std::endl;
-  if (parsed.fromSolvent()) {
-    out_stream << "Medium initialized from solvent built-in data." << std::endl;
-    Solvent solvent = parsed.solvent();
-    out_stream << solvent << std::endl;
-  }
-  out_stream << ".... Inside " << std::endl;
-  out_stream << *gf_i << std::endl;
-  out_stream << ".... Outside " << std::endl;
-  out_stream << *gf_o << std::endl;
-  out_stream << parsed.molecule() << std::endl;
+  size_t size = context_.getCavitySize();
 
   // Form vector with electrostatic potential
+  // First compute the potential from the classical point multipoles distribution
+  // then add the one from the molecule
   TIMER_ON("Computing MEP");
-  Eigen::VectorXd mep = computeMEP(parsed.molecule(), cavity->elements());
+  Eigen::VectorXd mep = Eigen::VectorXd::Zero(size);
+  if (input.MEPfromMolecule())
+    mep += computeMEP(context_.molecule(), context_.getCenters());
   TIMER_OFF("Computing MEP");
+  context_.setSurfaceFunction(mep.size(), mep.data(), "MEP");
   // Compute apparent surface charge
+  int irrep = 0;
   TIMER_ON("Computing ASC");
-  Eigen::VectorXd asc = solver->computeCharge(mep);
+  context_.computeASC("MEP", "ASC", irrep);
+  Eigen::VectorXd asc(size);
+  context_.getSurfaceFunction(asc.size(), asc.data(), "ASC");
+  context_.computeResponseASC("MEP", "RspASC", irrep);
+  Eigen::VectorXd rsp_asc(size);
+  context_.getSurfaceFunction(rsp_asc.size(), rsp_asc.data(), "RspASC");
   TIMER_OFF("Computing ASC");
   // Compute energy and print it out
-  out_stream << "Solvation energy = " << 0.5 * (asc.dot(mep)) << std::endl;
-  out_stream << "DONE!" << std::endl;
+  pcmsolver_out << "Solvation energy = "
+                << context_.computePolarizationEnergy("MEP", "ASC") << std::endl;
+  pcmsolver_out << "DONE!" << std::endl;
 
-  out_stream.close();
-
-  // Clean-up
-  TIMER_ON("Clean-up");
-  delete cavity;
-  delete gf_i;
-  delete gf_o;
-  delete op;
-  delete solver;
-  TIMER_OFF("Clean-up");
-
+  pcmsolver_out.close();
   // Write timings out
-  TIMER_DONE("pcmsolver.timer.dat");
+  context_.writeTimings();
 
   return EXIT_SUCCESS;
 }
 
-void host_writer(const char * /* message */, PCMSolverIndex /* message_length */) {}
+void host_writer(const char * message, int /* message_length */) {
+  pcmsolver_out << std::string(message) << std::endl;
+}
+
+std::string remove_extension(const std::string & filename) {
+  size_t lastdot = filename.find_last_of(".");
+  if (lastdot == std::string::npos)
+    return filename;
+  return filename.substr(0, lastdot);
+}